Split tests and added test identify colorspace. Updated Readme. Minor clean up.
